Make MainViewModel a singleton and move dialogs into MainWindow to enable WebView hotkey processing to support commands previously scoped to MainPage only
