+++ src/WinBlur.App/MainPage.xaml
@@ -5,16 +5,6 @@
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
       xmlns:vmodel="using:WinBlur.App.ViewModel"
       mc:Ignorable="d">
-
-    <Page.Resources>
-        <DataTemplate x:Key="FolderLabelItemTemplate"
-                      x:DataType="vmodel:FolderLabel">
-            <TextBlock Text="{x:Bind Title}"
-                       Margin="{x:Bind Depth, Converter={StaticResource DepthToMarginConverter}, ConverterParameter=20}"
-                       Style="{StaticResource BodyTextBlockStyle}" />
-        </DataTemplate>
-    </Page.Resources>
-
     <!--LayoutRoot is the root grid where all page content is placed-->
     <Grid x:Name="LayoutRoot">
         <VisualStateManager.VisualStateGroups>
@@ -258,126 +248,6 @@
                 Background="Transparent"
                 IsHitTestVisible="False"
                 Visibility="{x:Bind viewModel.TestModeHelper.TestMode, Mode=OneWay}" />
-
-        <ContentDialog x:Name="AddSiteDialog"
-                       Title="Add site"
-                       CloseButtonText="Cancel"
-                       PrimaryButtonText="Add Site"
-                       DefaultButton="Primary"
-                       PrimaryButtonClick="AddSiteDialog_PrimaryButtonClick">
-            <StackPanel Width="300" Margin="0,10,0,0">
-                <AutoSuggestBox x:Name="addSiteBox"
-                                PlaceholderText="Site name or feed URL"
-                                QueryIcon="Find"
-                                Margin="0,0,0,10"
-                                ItemsSource="{x:Bind viewModel.SiteAutoCompleteList, Mode=OneWay}"
-                                TextMemberPath="FeedURL"
-                                UpdateTextOnSelect="True"
-                                MaxSuggestionListHeight="170"
-                                TextChanged="addSiteBox_TextChanged"
-                                QuerySubmitted="addSiteBox_QuerySubmitted">
-                    <AutoSuggestBox.ItemTemplate>
-                        <DataTemplate x:DataType="vmodel:SiteAutoCompleteEntry">
-                            <StackPanel Margin="0,5">
-                                <TextBlock Text="{x:Bind Title}"
-                                           Style="{StaticResource BaseTextBlockStyle}" />
-                                <TextBlock Text="{x:Bind FeedURL}"
-                                           Foreground="{ThemeResource SystemControlForegroundBaseMediumHighBrush}"
-                                           Style="{StaticResource CaptionTextBlockStyle}" />
-                            </StackPanel>
-                        </DataTemplate>
-                    </AutoSuggestBox.ItemTemplate>
-                </AutoSuggestBox>
-                <ComboBox x:Name="addSiteFolderPicker"
-                          PlaceholderText="Parent folder"
-                          Margin="0,0,0,20"
-                          HorizontalAlignment="Stretch"
-                          ItemsSource="{x:Bind viewModel.FolderList, Mode=OneWay}"
-                          ItemTemplate="{StaticResource FolderLabelItemTemplate}" />
-
-                <StackPanel x:Name="AddSiteLoadingPanel" Orientation="Horizontal" HorizontalAlignment="Center" Spacing="10" Visibility="Collapsed">
-                    <ProgressRing x:Name="AddSiteProgressRing"
-                                  Width="20"
-                                  Height="20" />
-                    <TextBlock Text="Adding site..." />
-                </StackPanel>
-                <TextBlock x:Name="AddSiteErrorText"
-                           Foreground="{StaticResource RedAccentColorBrush}"
-                           TextWrapping="WrapWholeWords"
-                           Visibility="Collapsed"/>
-            </StackPanel>
-        </ContentDialog>
-
-        <ContentDialog x:Name="AddFolderDialog"
-                       Title="Add folder"
-                       CloseButtonText="Cancel"
-                       PrimaryButtonText="Add Folder"
-                       DefaultButton="Primary"
-                       PrimaryButtonClick="AddFolderDialog_PrimaryButtonClick">
-            <StackPanel Width="300" Margin="0,10,0,0">
-                <TextBox x:Name="addFolderBox"
-                         PlaceholderText="Folder name"
-                         Margin="0,0,0,10" />
-                <ComboBox x:Name="addFolderFolderPicker"
-                          PlaceholderText="Parent folder"
-                          Margin="0,0,0,10"
-                          HorizontalAlignment="Stretch"
-                          ItemsSource="{x:Bind viewModel.FolderList, Mode=OneWay}"
-                          ItemTemplate="{StaticResource FolderLabelItemTemplate}" />
-
-                <StackPanel x:Name="AddFolderLoadingPanel"
-                            Orientation="Horizontal"
-                            HorizontalAlignment="Center"
-                            Spacing="10"
-                            Visibility="Collapsed">
-                    <ProgressRing x:Name="AddFolderProgressRing"
-                                  Width="20"
-                                  Height="20" />
-                    <TextBlock Text="Adding folder..." />
-                </StackPanel>
-                <TextBlock x:Name="AddFolderErrorText"
-                           Foreground="{StaticResource RedAccentColorBrush}"
-                           TextWrapping="WrapWholeWords"
-                           Visibility="Collapsed"/>
-            </StackPanel>
-        </ContentDialog>
-
-        <ContentDialog x:Name="MarkAllAsReadDialog"
-                       Title="How many days back?"
-                       CloseButtonText="Cancel"
-                       DefaultButton="Primary"
-                       PrimaryButtonText="Submit"
-                       PrimaryButtonClick="MarkAllAsReadDialog_PrimaryButtonClick">
-            <StackPanel Width="300" Margin="0,10,0,0">
-                <TextBlock x:Name="markAllAsReadMessage"
-                           Text="Mark all stories older than 1 day as read"
-                           TextWrapping="WrapWholeWords"
-                           Margin="0,0,0,10" />
-                <Slider x:Name="markAllAsReadSlider"
-                        Minimum="0"
-                        Maximum="7"
-                        Value="1"
-                        TickFrequency="1"
-                        TickPlacement="Outside"
-                        SnapsTo="Ticks"
-                        Margin="0,0,0,10"
-                        ValueChanged="markAllAsReadSlider_ValueChanged" />
-                <StackPanel x:Name="MarkAllAsReadLoadingPanel"
-                            Orientation="Horizontal"
-                            HorizontalAlignment="Center"
-                            Spacing="10"
-                            Visibility="Collapsed">
-                    <ProgressRing x:Name="MarkAllAsReadProgressRing"
-                                  Width="20"
-                                  Height="20" />
-                    <TextBlock Text="Marking stories as read..." />
-                </StackPanel>
-                <TextBlock x:Name="MarkAllAsReadErrorText"
-                           Foreground="{StaticResource RedAccentColorBrush}"
-                           TextWrapping="WrapWholeWords"
-                           Visibility="Collapsed" />
-            </StackPanel>
-        </ContentDialog>
 
         <ContentDialog x:Name="DeleteDialog"
                        CloseButtonText="Cancel"
